adopts syscall for darwin platform

+++ stat_darwin.go
@@ -1,5 +1,5 @@
-//go:build dragonfly || freebsd || linux || netbsd || openbsd || solaris
-// +build dragonfly freebsd linux netbsd openbsd solaris
+//go:build darwin
+// +build darwin
 
 /*
  * knoxite
@@ -14,12 +14,12 @@ import "syscall"
 
 type statUnix syscall.Stat_t
 
-func (s statUnix) dev() uint64   { return s.Dev }
+func (s statUnix) dev() uint64   { return uint64(s.Dev) }
 func (s statUnix) ino() uint64   { return s.Ino }
-func (s statUnix) nlink() uint64 { return s.Nlink }
+func (s statUnix) nlink() uint64 { return uint64(s.Nlink) }
 func (s statUnix) uid() uint32   { return s.Uid }
 func (s statUnix) gid() uint32   { return s.Gid }
-func (s statUnix) rdev() uint64  { return s.Rdev }
+func (s statUnix) rdev() uint64  { return uint64(s.Rdev) }
 func (s statUnix) size() int64   { return s.Size }
 
 func toStatT(i interface{}) (statT, bool) {
